requirements.txt
# ── Core framework ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# ── LangChain + Groq ───────────────────────────────────────────────────────────────
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-experimental>=0.0.59
groq>=0.8.0

# ── Data ──────────────────────────────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
tabulate>=0.9.0

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit>=1.35.0
requests>=2.31.0

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
